change indeterminate logic (should be refactored)
